structure imp.

+++ src/main/java/ICar.java
@@ -2,15 +2,15 @@
 public interface ICar {
 	
 	// get default length (L0)
-	public float getL0();
+	public double getL0();
 	
 	// get default mass (M0)
-	public float getM0();
+	public double getM0();
 	
 	public void setCouplers(ICoupler[] couplers);
 	
 	// get total retardative force based on current published sensors' values
-	public float getF();
+	public double getF();
 	
 	// get a list of sensors installed in the car
 	public ISensor[] getSensors();
@@ -19,6 +19,6 @@ public interface ICar {
 	public ICoupler[] getCouplers();
 	
 	// read current GP value through the GP Pressure Sensor
-	public float getGPP();
+	public double getGPP();
 
 }
